[E-MEMO-RETIRE S3-1] MCP 메모 도구 제거 — chat 도구 보존 (#927)
* feat(mcp): E-MEMO-RETIRE S3-1 — MCP 메모 도구 제거, chat 도구 보존

- tools/memos.py 삭제
- tools/chat.py 신설 (send_chat_message, create_conversation, list_chat_messages 3개 보존)
- server.py: memos import → chat import, 메모 도구 7개 등록 제거


* test(mcp): E-MEMO-RETIRE S3-1 — test_contract 메모 도구 7종 제거 반영

- 도구 수 88 → 86
- EXPECTED_TOOLS에서 memo 7종 제거, chat (3)으로 업데이트
- test_project_id_not_in_schema, test_optional_fields, test_flat_schema에서 list_memos 참조 제거


* test(mcp): E-MEMO-RETIRE S3-1 — test_e2e_dev 메모 도구 제거 반영

- test_tools_list_88_tools → test_tools_list_86_tools (86으로 수정)
- READ_ONLY_TOOLS에서 sprintable_list_memos, sprintable_list_my_memos 제거


---------
