+++ backend/tests/mcp/test_e2e_dev.py
@@ -42,8 +42,6 @@ READ_ONLY_TOOLS = [
     ("sprintable_get_project_health", {}),
     ("sprintable_get_sprint_velocity_history", {}),
     ("sprintable_list_team_members", {}),
-    ("sprintable_list_memos", {}),
-    ("sprintable_list_my_memos", {}),
     ("sprintable_list_meetings", {}),
     ("sprintable_list_retro_sessions", {}),
     ("sprintable_list_audit_logs", {}),
@@ -56,14 +54,14 @@ READ_ONLY_TOOLS = [
 
 
 @pytest.mark.anyio
-async def test_tools_list_88_tools():
-    """tools/list 응답에서 88개 도구 전량 확인."""
+async def test_tools_list_86_tools():
+    """tools/list 응답에서 86개 도구 전량 확인."""
     async with stdio_client(_SERVER_PARAMS) as (read, write):
         async with ClientSession(read, write) as session:
             await session.initialize()
             result = await session.list_tools()
             tool_names = {t.name for t in result.tools}
-            assert len(tool_names) == 88, f"도구 수 불일치: {len(tool_names)}"
+            assert len(tool_names) == 86, f"도구 수 불일치: {len(tool_names)}"
 
 
 @pytest.mark.anyio
